+++ scripts/cases.py
@@ -1,6 +1,6 @@
-from fem import PostProcessor
-from fem import FEMSystem
-from truss import TrussSystem
+from core.fem import PostProcessor
+from core.fem import FEMSystem
+from core.truss import TrussSystem
 
 import numpy as np
 
@@ -8,9 +8,7 @@ import numpy as np
 LOAD_CASES = ["debug"]  # ["pullup_limit", "pullup_ultimate", "lateral_gust", "torsion"]
 
 
-# =========================
-# LOAD CASES (ENVELOPE)
-# =========================
+
 def create_load_case(case_name, zdiv, load_mag_base):
     """
     dof: 0=X, 1=Y, 2=Z
@@ -22,32 +20,32 @@ def create_load_case(case_name, zdiv, load_mag_base):
     n2 = top0 + 2
     n3 = top0 + 3
 
-    L = float(load_mag_base)
+    L = load_mag_base
 
     if case_name == "debug":
         Ftot = 1.0 * L
-        load_nodes = np.array([[n0, 1], [n3, 1]], dtype=int)
-        load_magnitudes = np.array([+0.5 * Ftot, +0.5 * Ftot], dtype=float)
+        load_nodes = np.array([[n0, 1], [n3, 1]])
+        load_magnitudes = np.array([+0.5 * Ftot, +0.5 * Ftot])
 
     elif case_name == "pullup_limit":
         Ftot = 6.0 * L
-        load_nodes = np.array([[n0, 1], [n3, 1]], dtype=int)
-        load_magnitudes = np.array([+0.5 * Ftot, +0.5 * Ftot], dtype=float)
+        load_nodes = np.array([[n0, 1], [n3, 1]])
+        load_magnitudes = np.array([+0.5 * Ftot, +0.5 * Ftot])
 
     elif case_name == "pullup_ultimate":
         Ftot = 1.5 * (6.0 * L)
-        load_nodes = np.array([[n0, 1], [n3, 1]], dtype=int)
-        load_magnitudes = np.array([+0.5 * Ftot, +0.5 * Ftot], dtype=float)
+        load_nodes = np.array([[n0, 1], [n3, 1]])
+        load_magnitudes = np.array([+0.5 * Ftot, +0.5 * Ftot])
 
     elif case_name == "lateral_gust":
         Fx = 0.5 * L  # 30N se L=60
-        load_nodes = np.array([[n0, 0], [n3, 0]], dtype=int)
-        load_magnitudes = np.array([-Fx, -Fx], dtype=float)
+        load_nodes = np.array([[n0, 0], [n3, 0]])
+        load_magnitudes = np.array([-Fx, -Fx])
 
     elif case_name == "torsion":
         Fx = 0.5 * L
-        load_nodes = np.array([[n0, 0], [n3, 0], [n1, 0], [n2, 0]], dtype=int)
-        load_magnitudes = np.array([-Fx, -Fx, +Fx, +Fx], dtype=float)
+        load_nodes = np.array([[n0, 0], [n3, 0], [n1, 0], [n2, 0]])
+        load_magnitudes = np.array([-Fx, -Fx, +Fx, +Fx])
 
     else:
         raise ValueError(
@@ -57,9 +55,7 @@ def create_load_case(case_name, zdiv, load_mag_base):
     return load_nodes, load_magnitudes
 
 
-# =========================
-# SOLVE DESIGN (ENVELOPE)
-# =========================
+
 def solve_case(
     system,
     case_name,
